refactor(webhooks): split WebhooksSection into focused components
Was 481 lines doing list rendering + create/edit dialog + deliveries
inspector + form-state plumbing all in one component. Splits into:

- webhook-types.ts (44 lines) -- shared types, default event list, date
  formatter. Co-located so the dialogs can import without pulling the
  parent.
- webhook-editor-dialog.tsx (281 lines) -- owns the create/edit form
  state and the one-time-secret reveal step.
- webhook-deliveries-dialog.tsx (135 lines) -- owns its own deliveries
  fetch, keyed off the `webhook` prop so opening for endpoint B after
  A never flashes the wrong history.
- webhooks-section.tsx (216 lines) -- list, header, refresh callback,
  delete confirmation, hands off to the two dialogs.

Plus a small WebhookRow sub-component inside webhooks-section to keep
the .map() body legible.

Net same line count, but each file is now under the giant-component
threshold and each has one job. No behavior change.

Resolves no-giant-component for this section.
